+++ i18n/template.ts
@@ -132,6 +132,7 @@ const table: Partial<typeof __table> = {
     "worldmap.selectafolder": "Select a folder to copy {0} file",
     "worldmap.tooltip.adjacencies": "Adjecencies",
     "worldmap.tooltip.category": "Category",
+    "worldmap.tooltip.claimby": "Claim by",
     "worldmap.tooltip.coastal": "Coastal",
     "worldmap.tooltip.continent": "Continent",
     "worldmap.tooltip.controller": "Controller",