+++ src/previewdef/worldmap/definitions.ts
@@ -66,7 +66,7 @@ export interface ProvinceDefinition {
     coastal: boolean;
     terrain: string;
     continent: number;
-    localisedName?: string;
+    localisedName: string | undefined;
     lineNumber?: number;
 }
 
@@ -115,6 +115,7 @@ export interface State extends Region, TokenInFile {
     controller: WithCondition<string>[]; // return the first matching country tag
     provinces: number[];
     cores: WithCondition<string>[];  // each item is a country tag with a condition, representing the core of the state
+    claimBy: WithCondition<string>[]; // each item is a country tag with a condition, representing the claim of the state
     impassable: boolean;
     victoryPoints: Record<number, number | undefined>;
     resources: Record<string, number | undefined>;
@@ -171,6 +172,7 @@ interface WarningRiver extends WarningSourceBase {
 export interface Country {
     tag: string;
     color: number;
+    localisedName: string | undefined;
     file: string;
 }